+++ CHANGELOG.md
@@ -4,6 +4,7 @@ unreleased
 - added support for model plot fn
 - added plotting for model type ::fastmath/ols 
   (equivalent to all 6 plots of R function 'plot.lm'  with default params)
+- added `explore-all` fn to plot all variables of a dataset  
 
 1.6.2
 - add inference targets to `r-model-matrix`